Fix quotient method for accurate division
Updated the quotient method to return a double value for accurate division results.

+++ projects/LabTask1.java
@@ -15,7 +15,7 @@ public class LabTask1 {
     }
     
     public static double quotient(int a, int b){
-        return a / b;
+        return (double)a / b;
     }
     
     public static void main(String[] args) {
